requirements.txt
Pyrogram==2.0.106
gTTS==2.5.1
qrcode==7.4.2
pyromod==3.1.5
requests==2.30.0
pytube==7.0.16
python-decouple
